+++ federation/callback.ts
@@ -1,6 +1,7 @@
 import type { NodeInfo } from "../nodeinfo/types.ts";
 import type { Actor } from "../vocab/actor.ts";
 import type { Activity, CryptographicKey } from "../vocab/mod.ts";
+import type { Object } from "../vocab/vocab.ts";
 import type { PageItems } from "./collection.ts";
 import type { RequestContext } from "./context.ts";
 
@@ -33,6 +34,23 @@ export type ActorKeyPairDispatcher<TContextData> = (
   contextData: TContextData,
   handle: string,
 ) => CryptoKeyPair | null | Promise<CryptoKeyPair | null>;
+
+/**
+ * A callback that dispatches an object.
+ *
+ * @typeParam TContextData The context data to pass to the {@link Context}.
+ * @typeParam TObject The type of object to dispatch.
+ * @typeParam TParam The parameter names of the requested URL.
+ * @since 0.7.0
+ */
+export type ObjectDispatcher<
+  TContextData,
+  TObject extends Object,
+  TParam extends string,
+> = (
+  context: RequestContext<TContextData>,
+  values: Record<TParam, string>,
+) => TObject | null | Promise<TObject | null>;
 
 /**
  * A callback that dispatches a collection.
@@ -117,6 +135,29 @@ export type OutboxErrorHandler = (
 export type AuthorizePredicate<TContextData> = (
   context: RequestContext<TContextData>,
   handle: string,
+  signedKey: CryptographicKey | null,
+  signedKeyOwner: Actor | null,
+) => boolean | Promise<boolean>;
+
+/**
+ * A callback that determines if a request is authorized or not.
+ *
+ * @typeParam TContextData The context data to pass to the {@link Context}.
+ * @typeParam TParam The parameter names of the requested URL.
+ * @param context The request context.
+ * @param values The parameters of the requested URL.
+ * @param signedKey The key that was used to sign the request, or `null` if
+ *                  the request was not signed or the signature was invalid.
+ * @param signedKeyOwner The actor that owns the key that was used to sign the
+ *                       request, or `null` if the request was not signed or the
+ *                       signature was invalid, or if the key is not associated
+ *                       with an actor.
+ * @returns `true` if the request is authorized, `false` otherwise.
+ * @since 0.7.0
+ */
+export type ObjectAuthorizePredicate<TContextData, TParam extends string> = (
+  context: RequestContext<TContextData>,
+  values: Record<TParam, string>,
   signedKey: CryptographicKey | null,
   signedKeyOwner: Actor | null,
 ) => boolean | Promise<boolean>;